Refactor WalletProvider and AppKitProviders structure; update import paths for WalletProvider in App and WalletConnect components
