+++ oryn_fi/src/providers/WalletProvider.tsx
@@ -3,7 +3,7 @@ import { WagmiProvider } from 'wagmi'
 import { QueryClientProvider, QueryClient } from '@tanstack/react-query'
 import { createAppKit } from '@reown/appkit/react'
 import { WagmiAdapter } from '@reown/appkit-adapter-wagmi'
-import { mainnet, arbitrum, polygon, optimism } from '@reown/appkit/networks'
+import { sepolia, rootstockTestnet, hederaTestnet, arbitrumSepolia, baseSepolia } from '@reown/appkit/networks'
 
 const projectId = import.meta.env.VITE_WALLETCONNECT_PROJECT_ID
 
@@ -12,7 +12,7 @@ if (!projectId) {
 }
 
 const wagmiAdapter = new WagmiAdapter({
-    networks: [mainnet, arbitrum, polygon, optimism],
+    networks: [sepolia, rootstockTestnet, hederaTestnet, arbitrumSepolia, baseSepolia],
     projectId,
     ssr: true,
 })
@@ -22,8 +22,8 @@ const queryClient = new QueryClient()
 export const appKit = createAppKit({
     adapters: [wagmiAdapter],
     projectId,
-    networks: [mainnet, arbitrum, polygon, optimism],
-    defaultNetwork: mainnet,
+    networks: [rootstockTestnet, hederaTestnet, sepolia, arbitrumSepolia, baseSepolia],
+    defaultNetwork: sepolia,
     features: {
         analytics: true
     }
